+++ scm/breaker/main/src/core/host/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod breaker;
-pub(crate) use breaker::DefaultHostBreaker;