+++ scm/breaker/main/src/core/mod.rs
@@ -3,4 +3,3 @@
 
 pub(crate) mod breaker;
 pub(crate) mod default;
-pub(crate) mod host;